Working blocg page with posts, now to fix css and add images again
