+++ styles/globals.css
@@ -1,4 +1,9 @@
-
+@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Roboto+Slab&display=swap");
+@font-face {
+  font-family: "Brand";
+  src: url("../public/Brand.woff");
+}
 * {
   padding: 0;
   margin: 0;
@@ -16,6 +21,7 @@ nav {
   align-items: center;
   justify-content: center;
   padding: 5rem;
+  font-family: "Brand";
 }
 
 h1 {
